Support for the absence of Hugo's FrontMatter (#7)
* feat: 🎸 add --hugo option

* docs: ✏️ --hugo option

* docs: ✏️ update help

* fix: 🐛 missing to signature translate_page

* fixed bug original file got deletet when without --hugo header, resolved conflict, add feature when --output flag enabled, output files with .md extension

* add --hugo --no-hugo flag selection, if no --hugo/--no-hugo flag exist then default is hugo

* update readme.md exmaple.md

---------

+++ mdt.py
@@ -22,7 +22,7 @@ def get_dest_file_path(filename, from_, to, output):
         dest_file_path = str(dest_file_path)
     else:
         md_file_name = Path(filename).name
-        dest_file_path = os.path.join(output_directory, md_file_name + '.md' if to == const.LANG_EN else md_file_name + '.' + to + '.md')
+        dest_file_path = os.path.join(output_directory, md_file_name + '.md' if to == const.LANG_EN else md_file_name + '.md')
         dest_file_path = re.sub(r"^./|/(\./)+", "/", dest_file_path)
 
     if dest_file_path.endswith('.' + const.LANG_EN + '.' + to + '.md'):
@@ -31,24 +31,29 @@ def get_dest_file_path(filename, from_, to, output):
     return str(dest_file_path)
 
 
-def translate_page(filename, from_, to, output, debug):
+def translate_page(filename, from_, to, is_hugo, output, debug):
     src_file_path = filename + '.md' if from_ == const.LANG_EN else filename + '.' + from_ + '.md'
+
     temp_file_path = filename + '.temp.md'
     dest_file_path = get_dest_file_path(filename, from_, to, output)
 
     # md 前処理 (hugo header の分離と翻訳)
-    translated_header_yaml_data = processing.mdProcessingBeforeTranslation(
-        src_file_path, temp_file_path, from_, to)
-
-    # markdown to json
-    remark.mdToJson(temp_file_path)
+    translated_header_yaml_data = ""
+    if is_hugo:
+        translated_header_yaml_data = processing.mdProcessingBeforeTranslation(
+            src_file_path, temp_file_path, from_, to)
+        # markdown to json
+        remark.mdToJson(temp_file_path)
+        # delete temporary md file
+        subprocess.run('rm ' + temp_file_path, shell=True)
+    else:
+        translated_header_yaml_data = ""
+        # markdown to json
+        remark.mdToJson(src_file_path)
 
     # 検証用 json 1
     if debug:
         shutil.copyfile("output.json", "ast-from-md.json")
-
-    # delete temporary md file
-    subprocess.run('rm ' + temp_file_path, shell=True)
 
     # json 前処理 (語順対応のための置き換え)
     ast, lookup_table = processing.jsonProcessingBeforeTranslation()
@@ -86,20 +91,20 @@ def mutate_path(ctx, param, value):
 @click.option('--path', help='directory or file path where you want to translate', required=True,
               callback=mutate_path, type=click.Path(exists=True))
 @click.option('-r', '--recursive', is_flag=True, help='translate recursively for subdirectories', show_default=True)
-@click.option('--from', 'from_', help='source language', default=const.LANG_EN, show_default=True,
-              type=click.Choice(const.LANGUAGE_LIST))
+@click.option('--hugo/--no-hugo',  default=True, help='translate with hugo front matter, which is metadata about hugo site', show_default=True)
+@click.option('--from', 'from_', help='source language', default=const.LANG_EN, show_default=True, type=click.Choice(const.LANGUAGE_LIST))
 @click.option('--to', help='post-translation language', default=const.LANG_JA, show_default=True,
               type=click.Choice(const.LANGUAGE_LIST))
 @click.option('--output', help='directory where you want to output the translated contents', default="./",
               show_default=True, type=click.Path(exists=True))
 @click.option('--debug', is_flag=True, help='output some ast files for debug', show_default=True)
-def run(path, recursive, from_, to, output, debug):
+def run(path, recursive, hugo, from_, to, output, debug):
+    is_hugo = hugo
+
     if path.endswith(".md"):
-        if (path.endswith("." + from_ + ".md") or (
-                    from_ == const.LANG_EN and "." not in path.split("/")[-1].rstrip(".md"))):
-            translate_page(
-                re.sub('\.md$', '', path) if from_ == const.LANG_EN else re.sub('\.' + from_ + '.md', '', path),
-                from_, to, output, debug)
+        if (path.endswith("." + from_ + ".md") or (from_ == const.LANG_EN and "." not in path.split("/")[-1].rstrip(".md"))):
+            translate_page(re.sub('\.md$', '', path) if from_ == const.LANG_EN else re.sub('\.' + from_ + '.md', '', path), from_, to, is_hugo, output, debug)
+
     else:
         if recursive:
             for current_dir, dirs, files in os.walk(path):
@@ -108,7 +113,7 @@ def run(path, recursive, from_, to, output, debug):
                             from_ == const.LANG_EN and "." not in filename.rstrip(".md"))):
                         click.echo("translate " + os.path.join(current_dir, re.sub('\.md$', '', filename)))
                         translate_page(
-                            re.sub('\.md$', '', os.path.join(current_dir, filename)) if from_ == const.LANG_EN else re.sub('\.md$', '',  os.path.join(current_dir, filename)), from_, to, output, debug)
+                            re.sub('\.md$', '', os.path.join(current_dir, filename)) if from_ == const.LANG_EN else re.sub('\.md$', '',  os.path.join(current_dir, filename)), from_, to, is_hugo, output, debug)
 
         else:
             for filename in os.listdir(path):
@@ -117,7 +122,6 @@ def run(path, recursive, from_, to, output, debug):
                             from_ == const.LANG_EN and "." not in filename.split("/")[-1].rstrip(".md"))):
                         click.echo("translate " + os.path.join(path, re.sub('\.md$', '', filename)))
                         translate_page(
-                            re.sub('\.md$', '', os.path.join(current_dir, filename)) if from_ == const.LANG_EN else re.sub('\.md$', '',  os.path.join(current_dir, filename)), from_, to, output, debug)
-
+                            re.sub('\.md$', '', os.path.join(current_dir, filename)) if from_ == const.LANG_EN else re.sub('\.md$', '',  os.path.join(current_dir, filename)), from_, to, is_hugo, output, debug)
 if __name__ == '__main__':
     run()
